# Frontend/sssklv2-openapi.json
{
    "openapi": "3.1.1",
    "info": {
        "title": "SSSKLv2 | v1",
        "contact": {
            "name": "Scouting Wilo",
            "email": "webmaster@scoutingwilo.nl"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost:7193/"
        }
    ],
    "paths": {
...
        "/v1/ApplicationUser": {
            "get": {
                "tags": [
                    "ApplicationUser"
                ],
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "take",
                        "in": "query",
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int32",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ApplicationUserDto"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ApplicationUserDto"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ApplicationUserDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/ApplicationUser/{id}": {
            "get": {
                "tags": [
                    "ApplicationUser"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicationUserDetailedDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicationUserDetailedDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicationUserDetailedDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "ApplicationUser"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "null"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ApplicationUserUpdateDto"
                                    }
                                ]
                            }
                        },
                        "text/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "null"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ApplicationUserUpdateDto"
                                    }
                                ]
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "null"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ApplicationUserUpdateDto"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ApplicationUser"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/ApplicationUser/by-username/{username}": {
            "get": {
                "tags": [
                    "ApplicationUser"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicationUserDetailedDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicationUserDetailedDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicationUserDetailedDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/ApplicationUser/me": {
            "get": {
                "tags": [
                    "ApplicationUser"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicationUserDetailedDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicationUserDetailedDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicationUserDetailedDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "ApplicationUser"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "null"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ApplicationUserSelfUpdateDto"
                                    }
                                ]
                            }
                        },
                        "text/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "null"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ApplicationUserSelfUpdateDto"
                                    }
                                ]
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "null"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ApplicationUserSelfUpdateDto"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ApplicationUser"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApplicationUserDetailedDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "userName": {
                        "type": "string"
                    },
                    "email": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "emailConfirmed": {
                        "type": "boolean"
                    },
                    "phoneNumber": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "phoneNumberConfirmed": {
                        "type": "boolean"
                    },
                    "name": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "surname": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "saldo": {
                        "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
                        "type": [
                            "number",
                            "string"
                        ],
                        "format": "double"
                    },
                    "lastOrdered": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "profilePictureBase64": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "ApplicationUserDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "userName": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "saldo": {
                        "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
                        "type": [
                            "number",
                            "string"
                        ],
                        "format": "double"
                    },
                    "lastOrdered": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ApplicationUserSelfUpdateDto": {
                "type": "object",
                "properties": {
                    "phoneNumber": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "name": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "surname": {
                        "type": [
                            "null",
                            "string"
                        ]
                    }
                }
            },
            "ApplicationUserUpdateDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "userName": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "email": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "emailConfirmed": {
                        "type": "boolean"
                    },
                    "phoneNumber": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "phoneNumberConfirmed": {
                        "type": "boolean"
                    },
                    "name": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "surname": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "password": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "roles": {
                        "type": [
                            "null",
                            "array"
                        ],
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
        }
    },
...
}